20230703: Tested benchmark gen. and sim.

+++ run/compare_dyn.sh
@@ -1,25 +1,13 @@
 #!/bin/bash
+# set -x 
 
 # 设置默认值为 200 20 360
 start=${1:-200}
 step=${2:-20}
 end=${3:-360}
 cfg=${4:-"heavy"}
-
-# read -p "请输入内容（light/heavy）：" cfg
-
-if [ "$cfg" = "light" ]; then
-    fn="profiling_light.csv"
-elif [ "$cfg" = "medium" ]; then
-    fn="profiling_medium.csv"
-elif [ "$cfg" = "heavy" ]; then
-    fn="profiling.csv"
-else
-    echo "无效的输入"
-    exit 1
-fi
-
-# echo "文件名为：$filename"
+p_fn=${5:-"allocator_agent.py"}
+PY_ARGS=${@:6}
 
 # 循环
 for x in $(seq "$start" "$step" "$end"); do
@@ -30,14 +18,16 @@ for x in $(seq "$start" "$step" "$end"); do
     mkdir -p "$dir_path"
     fi
 
-    rm "log/$cfg/$x/dyn_${x}_jitter_dis.log.txt"
-    rm "log/$cfg/$x/dyn_${x}_jitter_en.log.txt"
-
-    
+    if [ -f "log/$cfg/$x/dyn_${x}_jitter_dis.log.txt" ]; then
+        rm "log/$cfg/$x/dyn_${x}_jitter_dis.log.txt"
+    fi
+    if [ -f "log/$cfg/$x/dyn_${x}_jitter_en.log.txt" ]; then
+        rm "log/$cfg/$x/dyn_${x}_jitter_en.log.txt"
+    fi
     # 目录已经存在或者已经创建成功，接下来就可以进行其他的操作
 
-    nohup python allocator_agent.py --test_case dynamic --num_cores $x --n_p 3 --profiling_filename $fn > log/$cfg/$x/dyn_${x}_jitter_dis.log.txt 2>&1 &
-    nohup python allocator_agent.py --test_case dynamic --jitter_sim_en --file_suffix var_0.2 --num_cores $x --n_p 3 --profiling_filename $fn > log/$cfg/$x/dyn_${x}_jitter_en.log.txt 2>&1 &
+    nohup python $p_fn --test_case dynamic --num_cores $x --n_p 3 ${PY_ARGS} > log/$cfg/$x/dyn_${x}_jitter_dis.log.txt 2>&1 &
+    nohup python $p_fn --test_case dynamic --jitter_sim_en --file_suffix var_0.2 --num_cores $x --n_p 3 ${PY_ARGS} > log/$cfg/$x/dyn_${x}_jitter_en.log.txt 2>&1 &
 done
 wait
